refactor: unify four duplicated wire-contract types
Four response shapes were declared twice — once server-side and once as a
hand-copied mirror in the web bundle — so the two could (and did) drift.
The repo already has the fix pattern: read-side DTOs live in
`api/src/views/types.ts`, and web re-exports them through
`@beevibe/api/views/types` via thin `web/lib/types/*` files. These four
had never been pulled into it.

- RuntimePanelEntry / DaemonPanelEntry / RuntimesListResponse: the web copy
  had already diverged — `cli_version`/`last_heartbeat`/`last_seen_at` typed
  as `?: string` where the server sends `string | null`, and `device_name`
  as optional where the schema is NOT NULL. The type lie surfaced the moment
  the copies were unified (two compile errors, now fixed).
- WorkProductDetail: web inlined the literal union for `type`; now sourced
  from core's `WorkProductType`.
- FindRepoCandidate / FindRepoSource: web carried a comment admitting the
  duplication; `CandidateSource` in find-repo.ts is renamed to the shared
  `FindRepoSource`.
- ReferencedRepo: web copy now re-exports core's canonical declaration.

Canonical declarations move to `views/types.ts`; the server modules that
owned them re-export for their existing importers. On the web side, the two
runtime-picker surfaces (agent settings + welcome walkthrough) held a 5th and
6th copy of the same flattened runtime-option shape — extracted to
`web/lib/runtime-options.ts`, dropping the now-dead `?? external_id`
fallbacks. All re-exports are type-only, so nothing new reaches the browser
bundle (verified: web build clean).

Net -59 lines. typecheck, lint (0 errors), and build green across the
monorepo; web tests pass (203). DB- and key-gated suites fail only on
missing Postgres/provider keys, as documented.
