+++ packages/api/src/routes/runtimes.ts
@@ -24,6 +24,11 @@ import type {
 } from "@beevibe/core";
 import { requireHuman } from "../auth/middleware.js";
 import type { DaemonHub } from "../runtime/hub.js";
+import type {
+  DaemonPanelEntry,
+  RuntimePanelEntry,
+  RuntimesListResponse,
+} from "../views/types.js";
 import { loadOwned, requireParam } from "./http-errors.js";
 
 export interface RuntimesRoutesDeps {
@@ -33,30 +38,12 @@ export interface RuntimesRoutesDeps {
   hub: DaemonHub;
 }
 
-export interface RuntimePanelEntry {
-  id: string;
-  cli: string;
-  cli_version: string | null;
-  last_heartbeat: string | null;
-  /** True iff a daemon WS client subscribed to this runtime is connected. */
-  online: boolean;
-  capabilities: Record<string, unknown>;
-  created_at: string;
-}
-
-export interface DaemonPanelEntry {
-  id: string;
-  device_name: string;
-  external_id: string;
-  last_seen_at: string | null;
-  created_at: string;
-  runtimes: RuntimePanelEntry[];
-}
-
-export interface RuntimesListResponse {
-  ok: true;
-  daemons: DaemonPanelEntry[];
-}
+/**
+ * The panel DTOs live in `views/types.ts` — the one module web imports
+ * (`@beevibe/api/views/types`), so the projection below and the UI that
+ * renders it read a single declaration. Re-exported for existing importers.
+ */
+export type { DaemonPanelEntry, RuntimePanelEntry, RuntimesListResponse };
 
 function projectRuntime(r: Runtime, online: boolean): RuntimePanelEntry {
   return {
