+++ packages/api/src/tools/find-repo.ts
@@ -40,6 +40,7 @@ import type {
   LearnedSkillRepository,
 } from "@beevibe/core";
 import type { AgentTool, AgentToolResult } from "./types.js";
+import type { FindRepoCandidate, FindRepoSource } from "../views/types.js";
 
 /**
  * Data layer URLs — all live in the public `beevibe-ai/beevibe-capabilities`
@@ -93,31 +94,13 @@ const FIND_REPO_SCHEMA = {
   additionalProperties: false,
 } as const;
 
-export type CandidateSource = "learned" | "community" | "trending" | "github";
-
-export interface FindRepoCandidate {
-  repo_url: string;
-  score: number;
-  /** Highest-precedence source (learned > community > trending > github). */
-  source: CandidateSource;
-  /** Every source that contributed to the score. Useful for debugging. */
-  sources: CandidateSource[];
-  /** Human-readable explanation of why this candidate scored. */
-  reason: string;
-  /** GitHub stars when available (best-effort enrich). */
-  stars?: number;
-  /** GitHub description when available. */
-  description?: string;
-  /** Programming language inferred from GitHub. */
-  language?: string;
-  /** Hydrated learned_skill row when source includes "learned". */
-  learned_skill?: {
-    id: string;
-    name: string;
-    goal_pattern: string;
-    invocation: string;
-  };
-}
+/**
+ * The candidate shape lives in `views/types.ts` — the one module web imports
+ * (`@beevibe/api/views/types`), so the scoring below and the capabilities
+ * page that renders it read a single declaration. Re-exported for existing
+ * importers; `FindRepoSource` was previously spelled `CandidateSource` here.
+ */
+export type { FindRepoCandidate, FindRepoSource };
 
 export interface FindRepoContext {
   agentId: string;
@@ -672,7 +655,7 @@ function isFilteredOutContent(description: string | null | undefined): boolean {
   return FILTERED_DESCRIPTION_PATTERN.test(description);
 }
 
-function pickPrimarySource(sources: CandidateSource[]): CandidateSource {
+function pickPrimarySource(sources: FindRepoSource[]): FindRepoSource {
   // Trust order: learned > trending > community > github. The user's
   // explicit preference is "prefer trendy ones", so trending wins the
   // label when a candidate has both trending velocity and community
