+++ packages/web/lib/types/find-repo.ts
@@ -1,0 +1,1 @@
+export type { FindRepoCandidate, FindRepoSource } from "@beevibe/api/views/types";
