+++ packages/web/lib/types/referenced-repos.ts
@@ -1,0 +1,1 @@
+export type { ReferencedRepo } from "@beevibe/core/services/referenced-repos";
