+++ packages/web/lib/types/runtimes.ts
@@ -1,0 +1,5 @@
+export type {
+  DaemonPanelEntry,
+  RuntimePanelEntry,
+  RuntimesListResponse,
+} from "@beevibe/api/views/types";
