+++ packages/web/lib/types/work-products.ts
@@ -1,0 +1,1 @@
+export type { WorkProductDetail } from "@beevibe/api/views/types";
